backend/pyproject.toml
[project]
name = "pidem-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "holidays>=0.92",
    "httpx>=0.28.1",
    "lightgbm>=4.6.0",
    "numpy>=2.4.3",
    "openai>=2.26.0",
    "openmeteo-requests>=1.7.5",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "prophet>=1.3.0",
    "pulp>=3.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "requests-cache>=1.3.1",
    "retry-requests>=2.0.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "sentence-transformers>=5.2.3",
    "shap>=0.51.0",
    "sqlalchemy>=2.0.48",
    "statsmodels>=0.14.6",
    "torch>=2.0.0",
    "uvicorn[standard]>=0.41.0",
    "xgboost>=3.2.0",
    "yfinance>=1.2.0",
    "psycopg[binary]>=3.3.3",
]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
